добавил задачи в модуль 6
